+++ build.R
@@ -3,6 +3,7 @@ knit("README.Rmd", "README.md")
 
 library(devtools)
 document("pkg")
+test("pkg")
 build("pkg")
 source("examples/examples.R")