+++ pkg/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(ggseas)
+
+test_check("ggseas")